[Fix] next/jest import error jest.config.js to jest.config.ts
