+++ package.json
@@ -39,19 +39,20 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-typescript": "^7.16.7",
     "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "12.1.2",
+    "@testing-library/user-event": "13.5.0",
     "@types/react": "17.0.38",
     "@types/react-toggle": "^4.0.3",
     "@typescript-eslint/eslint-plugin": "^5.10.0",
     "@typescript-eslint/parser": "^5.10.0",
+    "babel-jest": "27.4.5",
     "babel-plugin-module-resolver": "^4.1.0",
     "eslint": "8.7.0",
     "eslint-config-airbnb-typescript": "^16.1.0",
     "eslint-config-next": "12.0.8",
-    "typescript": "4.5.5",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
-    "babel-jest": "27.4.5",
     "eslint-plugin-testing-library": "5.0.1",
-    "jest": "27.4.5"
+    "jest": "27.4.5",
+    "ts-node": "^10.5.0",
+    "typescript": "4.5.5"
   }
 }
